requirements.txt
numpy
pandas
torch
umap-learn
polyscope
scikit-learn
matplotlib

